Fix ActivityType class: remove phantom methods
