+++ src/Services/CRM/Activity/Service/ActivityType.php
@@ -152,72 +152,4 @@ class ActivityType extends AbstractService
             $this->core->call('crm.activity.type.list')
         );
     }
-
-    /**
-     * Get task type fields.
-     *
-     * @link https://apidocs.bitrix24.ru/api-reference/crm/timeline/activities/types/crm-activity-type-add.html
-     *
-     * @throws BaseException
-     * @throws TransportException
-     */
-    #[ApiEndpointMetadata(
-        '',
-        '',
-        'Get task type fields.'
-    )]
-    public function getFields(): array
-    {
-        $list = $this->list()->getActivityTypes();
-        $fields = [];
-
-        if ($list !== [] && is_array($list)) {
-
-            $res = $list[0]->getData();
-
-            $i = 0;
-            foreach (array_keys($res) as $key) {
-                $fields[$i] = $key;
-                $i++;
-            }
-
-            unset($i);
-        }
-
-        return $fields;
-    }
-
-    public function getFieldsDescription(): array
-    {
-        $list = $this->list()->getActivityTypes();
-        $fields = [];
-
-        if ($list !== [] && is_array($list)) {
-
-            $res = $list[0]->getData();
-
-            $i = 0;
-            foreach ($res as $key => $value) {
-                $type = '';
-
-                if (is_string($value)) {
-                    $type = 'string';
-                } elseif (is_int($value)) {
-                    $type = 'int';
-                } elseif (is_bool($value)) {
-                    $type = 'bool';
-                } elseif (is_array($value)) {
-                    $type = 'array';
-                }
-
-                $fields[$key] = ['type' => $type];
-
-                $i++;
-            }
-
-            unset($i);
-        }
-
-        return $fields;
-    }
 }
